+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/InjectConstructorTransformer.kt
@@ -6,12 +6,12 @@ import dev.zacsweers.metro.compiler.Origins
 import dev.zacsweers.metro.compiler.Symbols
 import dev.zacsweers.metro.compiler.asName
 import dev.zacsweers.metro.compiler.generatedClass
+import dev.zacsweers.metro.compiler.ir.ClassFactory
 import dev.zacsweers.metro.compiler.ir.IrMetroContext
 import dev.zacsweers.metro.compiler.ir.assignConstructorParamsToFields
 import dev.zacsweers.metro.compiler.ir.createIrBuilder
 import dev.zacsweers.metro.compiler.ir.dispatchReceiverFor
 import dev.zacsweers.metro.compiler.ir.finalizeFakeOverride
-import dev.zacsweers.metro.compiler.ir.implementsProviderType
 import dev.zacsweers.metro.compiler.ir.irExprBodySafe
 import dev.zacsweers.metro.compiler.ir.irInvoke
 import dev.zacsweers.metro.compiler.ir.irTemporary
@@ -26,7 +26,6 @@ import dev.zacsweers.metro.compiler.ir.requireSimpleFunction
 import dev.zacsweers.metro.compiler.ir.thisReceiverOrFail
 import dev.zacsweers.metro.compiler.ir.typeAsProviderArgument
 import org.jetbrains.kotlin.descriptors.ClassKind
-import org.jetbrains.kotlin.ir.builders.IrBuilderWithScope
 import org.jetbrains.kotlin.ir.builders.irBlockBody
 import org.jetbrains.kotlin.ir.builders.irCallConstructor
 import org.jetbrains.kotlin.ir.builders.irGet
@@ -38,27 +37,20 @@ import org.jetbrains.kotlin.ir.declarations.IrConstructor
 import org.jetbrains.kotlin.ir.declarations.IrField
 import org.jetbrains.kotlin.ir.declarations.IrSimpleFunction
 import org.jetbrains.kotlin.ir.declarations.IrValueParameter
-import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.symbols.IrConstructorSymbol
-import org.jetbrains.kotlin.ir.symbols.IrFunctionSymbol
-import org.jetbrains.kotlin.ir.types.typeWith
 import org.jetbrains.kotlin.ir.util.callableId
 import org.jetbrains.kotlin.ir.util.classIdOrFail
 import org.jetbrains.kotlin.ir.util.companionObject
-import org.jetbrains.kotlin.ir.util.defaultType
 import org.jetbrains.kotlin.ir.util.file
 import org.jetbrains.kotlin.ir.util.functions
 import org.jetbrains.kotlin.ir.util.getAnnotation
 import org.jetbrains.kotlin.ir.util.getAnnotationStringValue
-import org.jetbrains.kotlin.ir.util.isFromJava
-import org.jetbrains.kotlin.ir.util.isObject
 import org.jetbrains.kotlin.ir.util.kotlinFqName
 import org.jetbrains.kotlin.ir.util.nestedClasses
 import org.jetbrains.kotlin.ir.util.nonDispatchParameters
 import org.jetbrains.kotlin.ir.util.packageFqName
 import org.jetbrains.kotlin.ir.util.parentAsClass
 import org.jetbrains.kotlin.ir.util.primaryConstructor
-import org.jetbrains.kotlin.ir.util.simpleFunctions
 import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.name.ClassId
 
@@ -67,20 +59,21 @@ internal class InjectConstructorTransformer(
   private val membersInjectorTransformer: MembersInjectorTransformer,
 ) : IrMetroContext by context {
 
-  private val generatedFactories = mutableMapOf<ClassId, ConstructorInjectedFactory>()
+  private val generatedFactories = mutableMapOf<ClassId, ClassFactory>()
 
   fun visitClass(declaration: IrClass) {
     val injectableConstructor =
       declaration.findInjectableConstructor(onlyUsePrimaryConstructor = false)
     if (injectableConstructor != null) {
-      getOrGenerateFactory(declaration, injectableConstructor)
+      getOrGenerateFactory(declaration, injectableConstructor, doNotErrorOnMissing = false)
     }
   }
 
   fun getOrGenerateFactory(
     declaration: IrClass,
-    targetConstructor: IrConstructor,
-  ): ConstructorInjectedFactory? {
+    previouslyFoundConstructor: IrConstructor?,
+    doNotErrorOnMissing: Boolean,
+  ): ClassFactory? {
     val injectedClassId: ClassId = declaration.classIdOrFail
     generatedFactories[injectedClassId]?.let {
       return it
@@ -115,18 +108,31 @@ internal class InjectConstructorTransformer(
     if (factoryCls == null) {
       if (isExternal) {
         if (options.enableDaggerRuntimeInterop) {
+          val targetConstructor =
+            previouslyFoundConstructor
+              ?: declaration.findInjectableConstructor(onlyUsePrimaryConstructor = false)
+              ?: return null
           // Look up where dagger would generate one
           val daggerFactoryClassId = injectedClassId.generatedClass("_Factory")
           val daggerFactoryClass = pluginContext.referenceClass(daggerFactoryClassId)?.owner
           if (daggerFactoryClass != null) {
-            val wrapper = ConstructorInjectedFactory.DaggerFactory(metroContext, daggerFactoryClass)
+            val wrapper =
+              ClassFactory.DaggerFactory(
+                metroContext,
+                daggerFactoryClass,
+                targetConstructor.parameters(metroContext),
+              )
             generatedFactories[injectedClassId] = wrapper
             return wrapper
           }
+        } else if (doNotErrorOnMissing) {
+          return null
         }
         declaration.reportError(
           "Could not find generated factory for '${declaration.kotlinFqName}' in upstream module where it's defined. Run the Metro compiler over that module too."
         )
+        return null
+      } else if (doNotErrorOnMissing) {
         return null
       } else {
         error(
@@ -139,7 +145,12 @@ internal class InjectConstructorTransformer(
     // TODO this doesn't work as expected in KMP, where things compiled in common are seen as
     //  external but no factory is found?
     if (isExternal) {
-      val wrapper = ConstructorInjectedFactory.MetroFactory(factoryCls)
+      val parameters =
+        factoryCls
+          .requireSimpleFunction(Symbols.StringNames.CONSTRUCTOR_FUNCTION)
+          .owner
+          .parameters(this)
+      val wrapper = ClassFactory.MetroFactory(factoryCls, parameters)
       generatedFactories[injectedClassId] = wrapper
       return wrapper
     }
@@ -147,6 +158,9 @@ internal class InjectConstructorTransformer(
     val injectors = membersInjectorTransformer.getOrGenerateAllInjectorsFor(declaration)
     val memberInjectParameters = injectors.flatMap { it.parameters.values.flatten() }
 
+    val targetConstructor =
+      previouslyFoundConstructor
+        ?: declaration.findInjectableConstructor(onlyUsePrimaryConstructor = false)!!
     val constructorParameters = targetConstructor.parameters(metroContext, factoryCls, declaration)
     val allParameters =
       buildList {
@@ -209,9 +223,19 @@ internal class InjectConstructorTransformer(
 
     possiblyImplementInvoke(declaration, constructorParameters)
 
+    // Generate a metadata-visible function that matches the signature of the target constructor
+    // This is used in downstream compilations to read the constructor's signature
+    val constructorFunction =
+      generateMetadataVisibleConstructorFunction(
+        context = metroContext,
+        factoryClass = factoryCls,
+        targetConstructor = targetConstructor,
+      )
+
     factoryCls.dumpToMetroLog()
 
-    val wrapper = ConstructorInjectedFactory.MetroFactory(factoryCls)
+    val wrapper =
+      ClassFactory.MetroFactory(factoryCls, constructorFunction.parameters(metroContext))
     generatedFactories[injectedClassId] = wrapper
     return wrapper
   }
@@ -448,86 +472,5 @@ internal class InjectConstructorTransformer(
         }
       }
     return newInstanceFunction
-  }
-
-  sealed interface ConstructorInjectedFactory {
-    val factoryClass: IrClass
-    val invokeFunctionSymbol: IrFunctionSymbol
-
-    fun IrBuilderWithScope.invokeCreateExpression(
-      computeArgs: IrBuilderWithScope.(createFunction: IrSimpleFunction) -> List<IrExpression?>
-    ): IrExpression
-
-    class MetroFactory(override val factoryClass: IrClass) : ConstructorInjectedFactory {
-      override val invokeFunctionSymbol: IrFunctionSymbol
-        get() = factoryClass.requireSimpleFunction(Symbols.StringNames.INVOKE)
-
-      override fun IrBuilderWithScope.invokeCreateExpression(
-        computeArgs: IrBuilderWithScope.(IrSimpleFunction) -> List<IrExpression?>
-      ): IrExpression {
-        // Invoke its factory's create() function
-        val creatorClass =
-          if (factoryClass.isObject) {
-            factoryClass
-          } else {
-            factoryClass.companionObject()!!
-          }
-        val createFunction = creatorClass.requireSimpleFunction(Symbols.StringNames.CREATE)
-        val args = computeArgs(createFunction.owner)
-        return irInvoke(
-          dispatchReceiver = irGetObject(creatorClass.symbol),
-          callee = createFunction,
-          args = args,
-          typeHint = factoryClass.typeWith(),
-        )
-      }
-    }
-
-    class DaggerFactory(
-      private val metroContext: IrMetroContext,
-      override val factoryClass: IrClass,
-    ) : ConstructorInjectedFactory {
-      override val invokeFunctionSymbol: IrFunctionSymbol
-        get() = factoryClass.requireSimpleFunction(Symbols.StringNames.GET)
-
-      override fun IrBuilderWithScope.invokeCreateExpression(
-        computeArgs: IrBuilderWithScope.(createFunction: IrSimpleFunction) -> List<IrExpression?>
-      ): IrExpression {
-        // Anvil may generate the factory
-        val isJava = factoryClass.isFromJava()
-        val creatorClass =
-          if (isJava || factoryClass.isObject) {
-            factoryClass
-          } else {
-            factoryClass.companionObject()!!
-          }
-        val createFunction =
-          creatorClass
-            .simpleFunctions()
-            .first {
-              it.name == Symbols.Names.create || it.name == Symbols.Names.createFactoryProvider
-            }
-            .symbol
-        val args = computeArgs(createFunction.owner)
-        val createExpression =
-          irInvoke(
-            dispatchReceiver = if (isJava) null else irGetObject(creatorClass.symbol),
-            callee = createFunction,
-            args = args,
-            typeHint = factoryClass.typeWith(),
-          )
-
-        // Wrap in a metro provider if this is a provider
-        return if (factoryClass.defaultType.implementsProviderType(metroContext)) {
-          irInvoke(
-            extensionReceiver = createExpression,
-            callee = metroContext.symbols.daggerSymbols.asMetroProvider,
-            typeArgs = listOf(factoryClass.typeWith()),
-          )
-        } else {
-          createExpression
-        }
-      }
-    }
   }
 }